work started on user services

+++ recat/src/main/java/com/code/recat/user/UserRepository.java
@@ -1,6 +1,5 @@
 package com.code.recat.user;
 
-import com.code.recat.dto.UserDTO;
 import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     @Transactional
     Optional<User> findByEmail(String email);
